Fence can be called any time, even if execute() has not been called. We could parsec_context_start() if it has not been called when reaching fence(), the parsec_context_wait() in fence() would then rely on the global semantics for the termination detection... However, that termination detection might be less efficient than a simple MPI_Barrier().
If a user called execute() on some ranks but not all (which is incorrect),
the context_start/context_wait approach would be more tolerating, and
would make the code work. To the cost of a suboptimal fence() for correct
codes...

I opted for efficiency, but incorrect codes will result in a deadlock
in the fence().