(version of the commit without reformatting the code)
find_copy_in_task: we cannot find the copy if either the copy is NULL (new data, can happen in invoke()), or the caller task is NULL (can happen in invoke()).

find_copy_in_task can return NULL in the previous cases, in which case we need to create a new copy from the invoke() data.