SPMM: scramble the hash using a randomly chosen prime number
The default keymap uses the hash for process selection.
The indeces in Eigen are the index of the tile's first element
in the matrix so the key directly depends on the tile size.
If the tile size is a multiple of the number of processes all
tasks end up on one process... We better scramble the hash!

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>