Add support for generic callables
Instead of inferring constness from the callable's signature
we have to rely on input edge type information.
This adds ttg::make_const that can be used to mark an input edge const:

```
ttg::Edge<int, int> in;
ttg::make_tt([](const KeyT& key, auto& val, auto& out){...},
             ttg::edges(ttg::make_const(in)));
```

Even if the callable parameter `val` was not templated (using auto),
the application still has to mark the edge as const, since we have
no way to inspect any of the parameters of a generic callable.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>