More fun with invoke()
Invoke is special in that it can trigger a task that has no inputs.
Instead of special treatment of that case, we just add a virtual
control input if the task has no regular inputs. This also
cleared up some inconsistencies between the parsec and madness backend
in the way remote invoke() calls are handled.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>