Draft of aggregators
Aggregators are meant to replace reduction terminals where the sole purpose
is to collect a number of inputs before passing them into the
task. An aggregator is created by wrapping the input type in a
ttg::aggregator, which will cause values to be aggregated until either
the target count is reached (if provided) or the stream size has been reached.
The aggregator is then passed directly into the task and can be iterated over.

In combination with make_tt, an aggregator is created using
`ttg::make_aggregator(inedge)` or `ttg::make_aggregator(inedge, target)`.
These calls will return a specialized Edge that allows to check at
compile-time and create aggregators in the backend.

Example:

```
ttg::Edge<int, int> edge;
auto tt = make_tt(
  [](const int& key, ttg::aggregator<int>& agg){
    for (auto&& v : agg) {
       ...
    }
  },
  ttg::edges(ttg::make_aggregator(edge)), ...);
```

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>