Implement simple key ranges for broadcast
Problem: broadcasting a range of keys currently requires application
to populate the full range into a container and pass it to ttg::broadcast.
These keys then have to be processed and send to peers, leading to
large messages (in addition to the actual data transfer).

Many (dense) applications send a range of keys that can be expressed
as a linear function in an N-dimensional space. With key ranges,
application can express the start and end of a range as well as
a step size for iterating through the range (a vector in the N dimensional
space). Some applications may require splitting the range, i.e.,
providing a list or ranges (e.g., the Floyd-Warshall example).

These ranges can be serialized and send to peers instead of the fully
populated list of keys. Such compact representation of the key range
is cheaper to transfer and store.

An example for using key ranges:

```
auto range1_low = ttg::make_keyrange(Key(I, 0, K), /* start */
                                     Key(I, J, K), /* end, i.e., last + 1 */
                                     Key(0, 1, 0)); /* step vector */
```

It is possible to define a different type for the step vector by defining
`Key::difference_type` or by specializing the `ttg::key_diff_type` trait.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>