Allow TTs to defer a writer if there are readers
This feature is useful if there is sufficient concurrency in the application,
with multiple consumers and one subsequent producer on a data.
By seting defer_writer(), the application can signal that it is ok
to defer the next writer on the data until all readers completed,
avoiding the data to be copied.
This may lead to a deadlock though, if the execution of the readers
depends on the writer. This feature thus has to be used cautiously.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>