PaRSEC backend: offload reductions to dedicated task
When we encounter a streaming terminal we schedule a dedicated
reducer task. Subsequent inputs are put into a LIFO. Once the
reducer task runs it will process all available inputs for that
terminal but will not block waiting for all inputs. In the
worst case, we create N tasks (one for each input), likely less.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>