Add debugging capabilities for POTRF benchmark
Set DEBUG_TILES_VALUES=1 to enable norm2 check of tiles.

This also fixes issues with the parsec data wrapper. This needs some
more thought and might come back to haunt us if we try to use data with
extrinsic tracking information.

Also adds ttg::device::Device to query the ID and type of an instance of
the current execution space (e.g., the currently assigned GPU).

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>