Multi-GPU: Play nice with the PaRSEC accelerator
We need to a) push tiles back if they move from RW to RO tasks, and b)
create a fake task-class that contains all the flows we actually need
(i.e., that are part of to_device()). This way PaRSEC can select the
right device and we get all the transfers in multi-GPU settings.

This needs to be revisited in the PaRSEC backend. We should not need to
do the push-back and create a fake task-class to select the correct
device. That requires an overhaul of the device backend though...

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>