Towards device-enabled project function
Had to copy a bunch of headers to remove the SimpleTensor/FixedTensor
and adjust for potential compilation in CUDA context.
Does not yet compile because we cannot co_await from within
a function that is not a coroutine. Need to come up with a way to
switch between coroutines inside a task.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>