Discard temporary scratch data by setting the flow read-only
Set the flow to read-only at the end of the task for flows
that are marked as temporary. This allows PaRSEC to put the data
into a read-only queue that does not need to be pushed back to the host.
Temporary data does not need to be evicted, only released.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>