PaRSEC: check for universal peer access
Check whether thinks that all devices can access the memory of their peers.
If so, we do not need to push modified data back to the host as PaRSEC
will use D2D transfers instead of the host copy. We currently stick
to the owner-computes model so data will always be modified on the same
device and may only be read on a different device.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>