Select device copy and increment its readers
When we want to send from a device we need to increment the readers of
the device copy to make it's not evicted before we are done sending.
To avoid a race with the GPU manager, we have to lock the data_t
when looking for a suitable device copy.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>